chore(backward): backward compat data targeted generation
