+++ utils/tfhe-backward-compat-data/crates/generate_0_8/Cargo.toml
@@ -1,0 +1,21 @@
+[package]
+name = "generate_0_8"
+edition = "2024"
+version = "0.9.0"
+license = "BSD-3-Clause-Clear"
+
+[dependencies]
+clap = { version = "4.5", features = ["derive"] }
+
+# TFHE-rs
+tfhe = { git = "https://github.com/zama-ai/tfhe-rs.git", tag = "tfhe-rs-0.8.3", features = [
+    "boolean",
+    "integer",
+    "shortint",
+    "x86_64-unix",
+    "zk-pok",
+    "experimental-force_fft_algo_dif4",
+] }
+tfhe-versionable = { git = "https://github.com/zama-ai/tfhe-rs.git", tag = "tfhe-rs-0.8.3" }
+
+tfhe-backward-compat-data = { path = "../.." }
